Remove direct association between user and book models
Before a user had direct assoication to books, through user id since
the first design had put into consideration that an author might be a
user of our application, which is not correct.

This association will exist when a user has borrowed a book. The
association will be clear and reasonable when it comes from borrowing.